Split theme manager package core
